app/requirements.txt
pytz
objectpath
prometheus_client
pyYAML